gdb: split the stub into a transport-free core and the headless driver
The packet semantics (breakpoints, watchpoints, register watches, monitor
commands, memory and register access, reverse execution, the LoadSeg
tracker) now live in gdbstub::core::GdbCore, which owns no socket and
borrows the Emulator per call, replying through CoreReply
(Packet/Resume/Disconnect/Kill). The blocking --gdb driver keeps only
transport concerns in gdbstub::headless: RSP framing, ack mode, interrupt
polling, continue/step driving, and console delivery as O packets.
Monitor and loadseg output accumulates as console data on the core for
whichever driver delivers it.

The wire-test fixtures (the LoadSeg ROM program and the minimal RSP
client) move to gdbstub::testkit so a windowed driver's tests can reuse
them. No behaviour change: all existing wire tests are unchanged, plus a
new unit test that the core buffers monitor output as data.
